Extract existing "Make Transfer" functionality
