+++ packages/desktop-client/src/components/accounts/Account.tsx
@@ -308,6 +308,7 @@ type AccountInternalProps = {
   hideFraction: boolean;
   accountsSyncing: string[];
   dispatch: AppDispatch;
+  onSetTransfer: ReturnType<typeof useTransactionBatchActions>['onSetTransfer'];
 };
 type AccountInternalState = {
   search: string;
@@ -1336,49 +1337,11 @@ class AccountInternal extends PureComponent<
   };
 
   onSetTransfer = async (ids: string[]) => {
-    const onConfirmTransfer = async (ids: string[]) => {
-      this.setState({ workingHard: true });
-
-      const payees = this.props.payees;
-
-      const { data: transactions } = await runQuery(
-        q('transactions')
-          .filter({ id: { $oneof: ids } })
-          .select('*'),
-      );
-      const [fromTrans, toTrans] = transactions;
-
-      if (transactions.length === 2 && validForTransfer(fromTrans, toTrans)) {
-        const fromPayee = payees.find(
-          p => p.transfer_acct === fromTrans.account,
-        );
-        const toPayee = payees.find(p => p.transfer_acct === toTrans.account);
-
-        const changes = {
-          updated: [
-            {
-              ...fromTrans,
-              payee: toPayee?.id,
-              transfer_id: toTrans.id,
-            },
-            {
-              ...toTrans,
-              payee: fromPayee?.id,
-              transfer_id: fromTrans.id,
-            },
-          ],
-        };
-
-        await send('transactions-batch-update', changes);
-      }
-
-      await this.refetchTransactions();
-    };
-
-    await this.checkForReconciledTransactions(
+    this.setState({ workingHard: true });
+    await this.props.onSetTransfer(
       ids,
-      'batchEditWithReconciled',
-      onConfirmTransfer,
+      this.props.payees,
+      this.refetchTransactions,
     );
   };
 
@@ -1923,6 +1886,7 @@ type AccountHackProps = Omit<
   | 'onBatchLinkSchedule'
   | 'onBatchUnlinkSchedule'
   | 'onBatchDelete'
+  | 'onSetTransfer'
 >;
 
 function AccountHack(props: AccountHackProps) {
@@ -1934,6 +1898,7 @@ function AccountHack(props: AccountHackProps) {
     onBatchLinkSchedule,
     onBatchUnlinkSchedule,
     onBatchDelete,
+    onSetTransfer,
   } = useTransactionBatchActions();
 
   return (
@@ -1945,6 +1910,7 @@ function AccountHack(props: AccountHackProps) {
       onBatchLinkSchedule={onBatchLinkSchedule}
       onBatchUnlinkSchedule={onBatchUnlinkSchedule}
       onBatchDelete={onBatchDelete}
+      onSetTransfer={onSetTransfer}
       {...props}
     />
   );
