+++ packages/desktop-client/src/hooks/useTransactionBatchActions.ts
@@ -12,12 +12,14 @@ import {
 } from 'loot-core/shared/transactions';
 import { applyChanges, type Diff } from 'loot-core/shared/util';
 import {
+  PayeeEntity,
   type AccountEntity,
   type ScheduleEntity,
   type TransactionEntity,
 } from 'loot-core/types/models';
 
 import { useDispatch } from '../redux';
+import { validForTransfer } from 'loot-core/client/transfer';
 
 type BatchEditProps = {
   name: keyof TransactionEntity;
@@ -408,11 +410,59 @@ export function useTransactionBatchActions() {
     }
   };
 
+  const onSetTransfer = async (
+    ids: string[],
+    payees: PayeeEntity[],
+    onSuccess: (ids: string[]) => void,
+  ) => {
+    const onConfirmTransfer = async (ids: string[]) => {
+      const { data: transactions } = await runQuery(
+        q('transactions')
+          .filter({ id: { $oneof: ids } })
+          .select('*'),
+      );
+      const [fromTrans, toTrans] = transactions;
+
+      if (transactions.length === 2 && validForTransfer(fromTrans, toTrans)) {
+        const fromPayee = payees.find(
+          p => p.transfer_acct === fromTrans.account,
+        );
+        const toPayee = payees.find(p => p.transfer_acct === toTrans.account);
+
+        const changes = {
+          updated: [
+            {
+              ...fromTrans,
+              payee: toPayee?.id,
+              transfer_id: toTrans.id,
+            },
+            {
+              ...toTrans,
+              payee: fromPayee?.id,
+              transfer_id: fromTrans.id,
+            },
+          ],
+        };
+
+        await send('transactions-batch-update', changes);
+      }
+
+      onSuccess?.(ids);
+    };
+
+    await checkForReconciledTransactions(
+      ids,
+      'batchEditWithReconciled',
+      onConfirmTransfer,
+    );
+  };
+
   return {
     onBatchEdit,
     onBatchDuplicate,
     onBatchDelete,
     onBatchLinkSchedule,
     onBatchUnlinkSchedule,
+    onSetTransfer,
   };
 }
